+++ Classes/Domain/Exception/AdminInfoMailNotSendException.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the TYPO3 extension feuserregistration.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ *  (c) 2023 Kevin Chileong Lee, info@wacon.de, WACON Internet GmbH
+ */
+
+ namespace Wacon\Feuserregistration\Domain\Exception;
+
+ class AdminInfoMailNotSendException extends \Exception
+ {
+    
+ }